requirements.txt
tweepy==3.5.0
networkx==2.1
python-dotenv==0.9.1